refactor(FCL-269): remove xml_tools
This has been deprecated for a while now; remove xml_tools entirely from the client

BREAKING CHANGE:

Remove the `xml_tools` module
